base array-annotations class
